fix: stabilize inversion export typing

+++ lib/ui/project_workflow/plots_panel.dart
@@ -617,19 +617,21 @@ class InversionPlotPanel extends StatelessWidget {
     final measurementSpots = <FlSpot>[];
     final predictedSpots = <FlSpot>[];
     for (var i = 0; i < summary.observedRho.length; i++) {
-      final observed = summary.observedRho[i];
+      final observed = summary.observedRho[i].toDouble();
       if (!observed.isFinite || observed <= 0) {
         continue;
       }
       final depth = i < summary.measurementDepthsM.length
-          ? summary.measurementDepthsM[i]
-          : (summary.measurementDepthsM.isEmpty ? 0 : summary.measurementDepthsM.last);
+          ? summary.measurementDepthsM[i].toDouble()
+          : (summary.measurementDepthsM.isEmpty
+              ? 0.0
+              : summary.measurementDepthsM.last.toDouble());
       measurementSpots.add(FlSpot(_log10(observed), -depth));
       final predicted = i < summary.predictedRho.length
-          ? summary.predictedRho[i]
+          ? summary.predictedRho[i].toDouble()
           : summary.predictedRho.isEmpty
               ? observed
-              : summary.predictedRho.last;
+              : summary.predictedRho.last.toDouble();
       if (predicted.isFinite && predicted > 0) {
         predictedSpots.add(FlSpot(_log10(predicted), -depth));
       }
